Unify all Noop implementation in same file
